Fix regression from commit 1f275736097f12d1975e557eb84f3b332b6f6f88
.. this caused createdelete to run out of files.

+++ kernel/include/fs.h
@@ -128,10 +128,10 @@ struct inode *namei(const char *);
 struct inode *namei_with_fd(int dirfd, const char *path);
 struct inode *nameiparent(const char *, char *);
 struct inode *nameiparent_with_fd(int dirfd, const char *path, char *name);
-ssize_t inode_read(struct inode *ip, char *, off_t off, uint64_t n)
+ssize_t inode_read(struct inode *ip, char *, off_t off, size_t n)
 	__must_hold(&ip->lock);
 void inode_stat(struct inode *ip, struct stat *) __must_hold(&ip->lock);
-ssize_t inode_write(struct inode *ip, char *, off_t off, uint64_t n)
+ssize_t inode_write(struct inode *ip, char *, off_t off, size_t n)
 	__must_hold(&ip->lock);
 #endif
 #endif
